+++ src/Overrides/Broadcast/BudBroadcastManagerOverride.php
@@ -3,13 +3,15 @@ declare(strict_types=1);
 
 namespace Sprout\Bud\Overrides;
 
+use Illuminate\Broadcasting\BroadcastManager;
+use Illuminate\Contracts\Container\Container;
 use Illuminate\Contracts\Foundation\Application;
-use Sprout\Bud\Overrides\Auth\BudAuthManager;
+use Sprout\Bud\Overrides\Broadcast\BudBroadcastManager;
 use Sprout\Core\Contracts\BootableServiceOverride;
 use Sprout\Core\Overrides\BaseOverride;
 use Sprout\Core\Sprout;
 
-final class AuthManagerOverride extends BaseOverride implements BootableServiceOverride
+final class BudBroadcastManagerOverride extends BaseOverride implements BootableServiceOverride
 {
     /**
      * Boot a service override
@@ -28,15 +30,15 @@ final class AuthManagerOverride extends BaseOverride implements BootableServiceO
     {
         $original = null;
 
-        // If the auth manager has already been resolved
-        if ($app->resolved('auth')) {
+        // If the broadcast manager has already been resolved
+        if ($app->resolved(BroadcastManager::class)) {
             // We'll grab the manager
-            $original = $app->make('auth');
+            $original = $app->make(BroadcastManager::class);
             // and then tell the container to forget it
-            $app->forgetInstance('auth');
+            $app->forgetInstance(BroadcastManager::class);
         }
 
-        // Bind a replacement auth manager to enable Bud features
-        $app->singleton('auth', fn (Application $app) => new BudAuthManager($app, $original));
+        // Bind a replacement broadcast manager to enable Bud features
+        $app->singleton(BroadcastManager::class, fn (Container $app) => new BudBroadcastManager($app, $original));
     }
 }
