+++ src/Overrides/Mailer/BudMailerOverride.php
@@ -17,7 +17,7 @@ use Sprout\Core\Sprout;
  *
  * @extends \Sprout\Bud\Overrides\BaseOverride<\Illuminate\Mail\MailManager>
  */
-final class MailerOverride extends BaseOverride
+final class BudMailerOverride extends BaseOverride
 {
     /**
      * Get the name of the service being overridden.
